refactor: faq section
