+++ src/components/pages/pricing/faq/item/index.js
@@ -1,0 +1,3 @@
+import Item from './item';
+
+export default Item;
